+++ .gitignore
@@ -1,3 +1,4 @@
 dist/
 generated/
 node_modules/
+.DS_Store
