+++ package.json
@@ -10,22 +10,25 @@
   "author": "Weston Pace",
   "license": "Apache-2.0",
   "dependencies": {
-    "bootstrap": "^5.1.3",
+    "bootstrap": "^5.2.2",
+    "d3": "^7.6.1",
+    "font-awesome": "^4.7.0",
     "style-loader": "^3.3.1",
     "substrait": "file:../substrait-js/substrait-0.20.0.tgz",
-    "webpack": "^5.72.1",
+    "webpack": "^5.75.0",
     "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.9.0"
+    "webpack-dev-server": "^4.11.1"
   },
   "devDependencies": {
-    "@alenaksu/json-viewer": "^1.0.0",
+    "@alenaksu/json-viewer": "^2.0.0",
     "css-loader": "^6.7.1",
-    "eslint": "^8.16.0",
+    "eslint": "^8.27.0",
     "eslint-config-google": "^0.14.0",
-    "jest": "^28.1.0",
-    "mermaid": "^9.1.1",
-    "protobufjs": "~6.11.3",
-    "sass": "^1.52.1",
-    "sass-loader": "^13.0.0"
+    "jest": "^29.3.1",
+    "mermaid": "^9.2.2",
+    "protobufjs": "~7.1.2",
+    "sass": "^1.56.1",
+    "sass-loader": "^13.2.0",
+    "webpack-cli": "^4.10.0"
   }
 }
