+++ src/lib.js
@@ -3,6 +3,8 @@ const mermaid = require('mermaid');
 const {substrait} = require('substrait');
 const {substraitToJson} = require('./substrait-to-json');
 const { substraitToMermaid } = require('./substrait-to-mermaid');
+const {SubstraitParser} = require('./substrait-parser')
+const {substraitD3, buildGraph, drawGraph} = require('./substrait-to-d3')
 
 mermaid.initialize({startOnLoad: true});
 
@@ -19,13 +21,10 @@ window.onload = function() {
         const plan = substrait.Plan.decode(new Uint8Array(reader.result));
         const planJson = substraitToJson(plan);
         document.getElementById('plan-json').data = JSON.parse(planJson);
-
-        const mermaidEl = document.getElementById('plan-mermaid');
-        const mermaidCb = (svg) => {
-          mermaidEl.innerHTML = svg;
-        };
-        const planMermaid = substraitToMermaid(plan);
-        mermaid.render('plan-mermaid-graph', planMermaid, mermaidCb);
+        const subplan = new SubstraitParser(plan).planToNode(plan);
+        const graph = buildGraph(subplan);
+        console.log("graph: ",graph);
+        drawGraph(graph['nodes'],graph['edges']);
       } catch (e) {
         dst = document.getElementById('plan-json');
         dst.data = {error: `Error parsing the plan: ${e}`};
